backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
recurly>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
